+++ common/src/org/nasdanika/common/WorkInfo.java
@@ -10,7 +10,7 @@ public interface WorkInfo {
 	/**
 	 * @return Total number of work units. The same size is used for both ``execute()`` and ``undo()``.
 	 */
-	long size();
+	double size();
 	
 	/**
 	 * @return Display name of the work.
